refactor(RoomView): move omnichannel tests into __tests__ and drop stale rid-keyed claims
The routing-config suites were colocated instead of living under __tests__, and its reducer imported actionsTypes twice. The room/thread screen suite and useSubscriptionUnreads still described a rid-keyed store that no longer exists.
