+++ app/ee/omnichannel/hooks/__tests__/useCanReturnQueue.test.tsx
@@ -1,12 +1,12 @@
 import { renderHook, waitFor } from '@testing-library/react-native';
 import { Provider } from 'react-redux';
 
-jest.mock('../../../lib/services/restApi', () => ({ getRoutingConfig: jest.fn() }));
+jest.mock('../../../../lib/services/restApi', () => ({ getRoutingConfig: jest.fn() }));
 
-import { getRoutingConfig } from '../../../lib/services/restApi';
-import routingConfigSaga from '../sagas/routingConfig';
-import { createRecordingStore, cancelSagaTasks } from '../../../lib/testUtils/sagaStore';
-import { useCanReturnQueue } from './useCanReturnQueue';
+import { getRoutingConfig } from '../../../../lib/services/restApi';
+import routingConfigSaga from '../../sagas/routingConfig';
+import { createRecordingStore, cancelSagaTasks } from '../../../../lib/testUtils/sagaStore';
+import { useCanReturnQueue } from '../useCanReturnQueue';
 import type { ReactNode } from 'react';
 
 const mockGetRoutingConfig = getRoutingConfig as jest.MockedFunction<typeof getRoutingConfig>;
